% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importbcrp.R
\name{importbcrp}
\alias{importbcrp}
\title{ImportBCRP}
\usage{
importbcrp(x, y, z)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
A function to compute the mean of a vector
}
\examples{
importbcrp('PN01205PM','2005-8','2020-9')
\dontrun{ mean(1:1e99) }
}
