% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importbcrp.R
\name{importbcrp}
\alias{importbcrp}
\title{importbcrpr}
\usage{
importbcrp(x, y, z)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
importbcrpr
}
\examples{
importbcrp('PN01205PM','2005-8','2020-9')
}
